Move tests to existing file
